requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
Django==3.2.8
django-crispy-forms==1.13.0
factory-boy==3.2.0
Faker==9.2.0
inflection==0.5.1
iniconfig==1.1.1
mysqlclient==2.0.3
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
